Move Manifesto Link (#13)
This PR proposes moving the manifesto link with the contact links. The
rationale is that our manifesto is more of a "permanent" page that
should not get buried at the bottom of the publications.

This is just a suggestion from my part - if you all don't think it works
the way I proposed, feel free to close the PR.